app/requirements.txt
google-genai=1.4.0
python=3.13.2
python-telegram-bot=21.11.1